+++ scripts/go.mod
@@ -1,0 +1,7 @@
+module github.com/pulumi/pulumi/scripts
+
+go 1.13
+
+require github.com/pulumi/pulumi/sdk v0.0.0-00010101000000-000000000000
+
+replace github.com/pulumi/pulumi/sdk => ../sdk
